이미지 저장 방식 변경(현제 게시판만 수정됨)

+++ Front/project_Full/src/js/pages/board/boardMU/useBoardMU.js
@@ -49,29 +49,21 @@ export function useBoard() {
     // ===== ===== ===== ===== ===== ===== ===== ===== =====
 
     // 첨푸파일 저장
-    const setFile = useCallback(async (target, fileList, bNum) => {
-
-        if (fileList === undefined || fileList.length === 0) { return }
+    const encodeFile = useCallback(async (target, file) => {
 
         const formData = new FormData();
-        fileList.forEach((file) => {
-            formData.append(`files`, file); // 각 이미지 파일을 FormData에 추가
-        });
+        formData.append('file', file);
 
-        return fetch(SERVER_URL + `setFile?target=${target}`, {
+        return fetch(SERVER_URL + `encodeFile?target=${target}&orgName=${file.name}`, {
             method: 'POST',
             body: formData
 
         }).then((response) => {
-            if (!response.ok) {
-                throw new Error(response.status);
-            }
+            if (!response.ok) { throw new Error(response.status); }
 
             return response.json();
 
-        }).catch((e) => {
-            console.log(e);
-        })
+        }).catch((e) => { console.log(e); })
     }, [])
 
     // ===== ===== ===== ===== ===== ===== ===== ===== =====
@@ -84,9 +76,9 @@ export function useBoard() {
         let path = '';
 
         if (boardNum === undefined) {
-            path = SERVER_URL + `setDetail?target=${target}&jwt=${jwt}&boardNum=${0}`;
+            path = SERVER_URL + `submitDetail?target=${target}&jwt=${jwt}&boardNum=${0}`;
         } else {
-            path = SERVER_URL + `setDetail?target=${target}&jwt=${jwt}&boardNum=${boardNum}`;
+            path = SERVER_URL + `submitDetail?target=${target}&jwt=${jwt}&boardNum=${boardNum}`;
         }
 
         return fetch(path, {
@@ -95,25 +87,19 @@ export function useBoard() {
             body: JSON.stringify(board)
 
         }).then((response) => {
-            if (!response.status === 401) {
-                alert('로그인이 필요합니다.');
-            } else if (!response.ok) {
-                throw new Error(response.status);
-            }
+            if (!response.ok) { throw new Error(response.status); }
 
             return response.json();
 
-        }).catch((e) => {
-            console.log(e);
-
-        })
+        }).catch((e) => { console.log(e); })
     }, [])
 
-    // 이미지 DB 등록
+    // 이미지 정보 등록 및 파일 저장 
     const submitFile = useCallback(async (target, fileList, boardNum) => {
 
         if (fileList === undefined || fileList.length === 0) {
 
+            alert('저장되었습니다.');
             navigate(`/boardList/${target}`);
             return;
         }
@@ -128,6 +114,7 @@ export function useBoard() {
                 throw new Error(response.status);
             }
 
+            alert('저장되었습니다.');
             navigate(`/boardList/${target}`);
 
         }).catch((e) => {
@@ -141,43 +128,31 @@ export function useBoard() {
     // 게시판 삭제
     const deleteBoard = useCallback(async (target, bNum) => {
 
-        if (!window.confirm('정말로 삭제하시겠습니까? 다시 복수할수 없습니다!')) { return; }
+        const jwt = sessionStorage.getItem('jwt');
 
+        if (!window.confirm('정말로 삭제하시겠습니까? 다시 복수할수 없습니다!')) { return; }
+        if (jwt === undefined || jwt === '') { alert('로그인이 필요합니다'); return; }
+
+        // ----- 저장된 이미지 정보와 파일 제거 -----
+        fetch(SERVER_URL + `deleteAllFile?target=${target}&boardNum=${bNum}&jwt=${jwt}`, {
+            method: 'DELETE'
+
+        }).then((res) => {
+            if (!res.json()) { throw new Error(res.status); }
+
+        }).catch((e) => { console.log(e); alert("삭제에 실패하였습니다."); })
+
+        // ----- 저장된 게시판 정보 제거 -----
         fetch(SERVER_URL + `boardDelete?target=${target}&boardNum=${bNum}`, {
             method: 'DELETE'
 
-        }).then((response) => {
-            if (response.status === 404) {
-                return undefined;
-            } else if (!response.ok) {
-                throw new Error(response.status);
-            }
+        }).then((res) => {
+            if (!res.ok) { throw new Error(res.status); }
 
             alert("삭제가 완료되었습니다.");
             navigate(`/boardList/${target}`);
 
-        }).catch((e) => {
-            console.log(e);
-            alert("삭제에 실패하였습니다.");
-        })
-    }, [])
-
-    // 이미지 첨부 취소
-    const deleteFile = useCallback(async (target, file) => {
-
-        return fetch(SERVER_URL + `deleteFile?target=${target}`, {
-            method: 'DELETE',
-            headers: { 'Content-Type': 'application/json' },
-            body: JSON.stringify(file)
-
-        }).then((response) => {
-            if (!response.ok) {
-                throw new Error(response.status);
-            }
-
-        }).catch((e) => {
-            console.log(e);
-        })
+        }).catch((e) => { console.log(e); alert("삭제에 실패하였습니다."); })
 
     }, [])
 
@@ -185,12 +160,11 @@ export function useBoard() {
         getDetail,
         getFile,
 
-        setFile,
+        encodeFile,
 
         submitDetail,
         submitFile,
 
-        deleteBoard,
-        deleteFile
+        deleteBoard
     }
 }
